requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
JSON-log-formatter==0.5.1
MarkupSafe==2.1.0
model-bakery==1.4.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
